build(swagger): compile UI only for dev builds
