+++ cmd/gomodel/main.go
@@ -12,7 +12,6 @@ import (
 	"syscall"
 	"time"
 
-	swaggerdocs "gomodel/cmd/gomodel/docs"
 	"gomodel/config"
 	"gomodel/internal/app"
 	"gomodel/internal/observability"
@@ -107,7 +106,7 @@ func main() {
 		slog.Error("failed to load config", "error", err)
 		os.Exit(1)
 	}
-	swaggerdocs.SwaggerInfo.BasePath = result.Config.Server.BasePath
+	configureSwaggerDocs(result.Config.Server.BasePath)
 
 	factory := providers.NewProviderFactory()
