+++ cmd/gomodel/swagger_disabled.go
@@ -1,0 +1,5 @@
+//go:build !swagger
+
+package main
+
+func configureSwaggerDocs(_ string) {}
