+++ cmd/gomodel/swagger_enabled.go
@@ -1,0 +1,9 @@
+//go:build swagger
+
+package main
+
+import swaggerdocs "gomodel/cmd/gomodel/docs"
+
+func configureSwaggerDocs(basePath string) {
+	swaggerdocs.SwaggerInfo.BasePath = basePath
+}
