+++ internal/server/swagger_disabled_test.go
@@ -1,0 +1,23 @@
+//go:build !swagger
+
+package server
+
+import (
+	"net/http"
+	"net/http/httptest"
+	"testing"
+)
+
+func TestSwaggerEndpoint_RequestedButNotAvailable(t *testing.T) {
+	mock := &mockProvider{}
+	srv := New(mock, &Config{SwaggerEnabled: true})
+
+	req := httptest.NewRequest(http.MethodGet, "/swagger/index.html", nil)
+	rec := httptest.NewRecorder()
+
+	srv.ServeHTTP(rec, req)
+
+	if rec.Code != http.StatusNotFound {
+		t.Errorf("expected status 404, got %d", rec.Code)
+	}
+}
